change DateRangePicker into functionnal component
