+++ package.json
@@ -20,8 +20,13 @@
   "dependencies": {
     "@cypress/react18": "^2.0.1",
     "@cypress/webpack-preprocessor": "^6.0.2",
+    "@emotion/react": "^11.13.5",
+    "@emotion/styled": "^11.13.5",
     "@geoman-io/leaflet-geoman-free": "^2.17.0",
     "@hello-pangea/dnd": "^16.5.0",
+    "@mui/material": "^6.1.8",
+    "@mui/x-date-pickers": "^7.22.3",
+    "@mui/x-date-pickers-pro": "^7.22.3",
     "cypress": "^13.15.0",
     "dotenv": "^8.6.0",
     "form-data": "^2.3.3",
